ip_quadruple/proxy: make port uint16_t
Make `port` member in these struct of type `uint16_t`.

CAVEAT: this change is visible to libcurl users in CURLINFO_PRIMARY_PORT
and CURLINFO_LOCAL_PORT when the connection was never really opened. Before
this change, -1 was returned as port number, now it is 0.

+++ lib/connect.h
@@ -74,7 +74,7 @@ curl_socket_t Curl_getconnectinfo(struct Curl_easy *data,
                                   struct connectdata **connp);
 
 bool Curl_addr2string(struct sockaddr *sa, curl_socklen_t salen,
-                      char *addr, int *port);
+                      char *addr, uint16_t *port);
 
 /*
  * Curl_conncontrol() marks the end of a connection/stream. The 'closeit'
